pyproject.toml
[tool.poetry]
name = "JAN23BDE_Itineraire_Vacances"
version = "0.1.0"
description = ""
authors = ["adrnbttr <adrnbttr@gmail.com>"]
maintainers = ["adrnbttr <adrnbttr@gmail.com>"]
license = ""
readme = "README.md"
homepage = ""
repository = ""
documentation = ""
keywords = []

[tool.poetry.dependencies]
python = "^3.9"
#python = "^3.10"
Flask = "^2.0.1"
dash = "^2.0.0"
dash-bootstrap-components = "^1.0.0"
pandas = "^1.3.1"
#scikit-learn = "^0.24.2"
Markdown = "^3.3.4"
python-dotenv = "^0.19.0"
Pygments = "^2.9.0"
dash-leaflet = "^0.1.23"
protobuf = "^3.20"
apache-airflow = "^2.5.1"
pymongo = "^4.3.3"
dash-flask-login = "^0.0.7"
bcrypt = "^4.0.1"
flask-bcrypt = "^1.0.1"
numpy = "^1.24.2"
torch = "^2.0.0"
torch-kmeans="^0.2.0"
selenium="^4.8.0"
#tensorboard = "^2.12.0"
#tensorboard-data-server = "^0.7.0"
#tensorboard-plugin-wit = "^1.8.1"
#dash-extensions = "^0.1.13"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
exclude = [
    "**/__pycache__",
]
typeCheckingMode = "basic"
